% Auto-generated: do not edit by hand
\name{dashIgv}

\alias{dashIgv}

\title{DashIgv component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashIgv(id=NULL, label=NULL, value=NULL, reference=NULL,
minimumBases=NULL, style=NULL, className=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{value}{Character. The value displayed in the input.}

\item{reference}{Named list. }

\item{minimumBases}{Named list. Minimum window size in base pairs when zooming in}

\item{style}{Named list. Generic style overrides on the plot div}

\item{className}{Character. className of the parent div}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

